requirements.txt
pandas 
requests 
tabula-py
apache-airflow